+++ aiocogeo/compression.py
