+++ aiocogeo/constants.py
