+++ aiocogeo/errors.py
